refactor: move parseRule to separate module
